+++ packages/react/src/components/InteractiveList/src/InteractiveList.tsx
@@ -13,6 +13,11 @@ import {
   InteractiveListCheckboxItem,
   InteractiveListRadioItem,
   InteractiveListToggleItem,
+  InteractiveListCheckboxItemSkeleton,
+  InteractiveListStructureSkeleton,
+  InteractiveListButtonItemSkeleton,
+  InteractiveListRadioItemSkeleton,
+  InteractiveListToggleItemSkeleton,
 } from "./components";
 
 const InteractiveList: React.FC<InteractiveListProps> &
@@ -28,10 +33,15 @@ const InteractiveList: React.FC<InteractiveListProps> &
 );
 
 InteractiveList.Structure = InteractiveListStructure;
+InteractiveList.StructureSkeleton = InteractiveListStructureSkeleton;
 InteractiveList.ButtonItem = InteractiveListButtonItem;
+InteractiveList.ButtonItemSkeleton = InteractiveListButtonItemSkeleton;
 InteractiveList.CheckboxItem = InteractiveListCheckboxItem;
+InteractiveList.CheckboxItemSkeleton = InteractiveListCheckboxItemSkeleton;
 InteractiveList.RadioItem = InteractiveListRadioItem;
+InteractiveList.RadioItemSkeleton = InteractiveListRadioItemSkeleton;
 InteractiveList.ToggleItem = InteractiveListToggleItem;
+InteractiveList.ToggleItemSkeleton = InteractiveListToggleItemSkeleton;
 
 InteractiveList.displayName = "InteractiveList";
 InteractiveList.Structure.displayName = "InteractiveList.Structure";
@@ -39,5 +49,15 @@ InteractiveList.ButtonItem.displayName = "InteractiveList.ButtonItem";
 InteractiveList.CheckboxItem.displayName = "InteractiveList.CheckboxItem";
 InteractiveList.RadioItem.displayName = "InteractiveList.RadioItem";
 InteractiveList.ToggleItem.displayName = "InteractiveList.ToggleItem";
+InteractiveList.StructureSkeleton.displayName =
+  "InteractiveList.StructureSkeleton";
+InteractiveList.ButtonItemSkeleton.displayName =
+  "InteractiveList.ButtonItemSkeleton";
+InteractiveList.CheckboxItemSkeleton.displayName =
+  "InteractiveList.CheckboxItemSkeleton";
+InteractiveList.RadioItemSkeleton.displayName =
+  "InteractiveList.RadioItemSkeleton";
+InteractiveList.ToggleItemSkeleton.displayName =
+  "InteractiveList.ToggleItemSkeleton";
 
 export { InteractiveList };