+++ packages/react/src/components/InteractiveList/src/components/InteractiveListButtonItemSkeleton/interactiveListButtonItemSkeleton.types.ts
@@ -1,0 +1,9 @@
+import { HTMLAttributes } from "react";
+import { InteractiveListStructureSkeletonProperties } from "../InteractiveListStructureSkeleton/interactiveListStructureSkeleton.types";
+
+export type InteractiveListButtonItemSkeletonProperties =
+  InteractiveListStructureSkeletonProperties;
+
+export type InteractiveListButtonItemSkeletonProps =
+  InteractiveListButtonItemSkeletonProperties &
+    Omit<HTMLAttributes<HTMLElement>, "color">;