+++ packages/react/src/components/InteractiveList/src/components/InteractiveListRadioItemSkeleton/interactiveListRadioItemSkeleton.types.ts
@@ -1,0 +1,9 @@
+import { HTMLAttributes } from "react";
+import { InteractiveListStructureSkeletonProperties } from "../InteractiveListStructureSkeleton/interactiveListStructureSkeleton.types";
+
+export type InteractiveListRadioItemSkeletonProperties =
+  InteractiveListStructureSkeletonProperties;
+
+export type InteractiveListRadioItemSkeletonProps =
+  InteractiveListRadioItemSkeletonProperties &
+    Omit<HTMLAttributes<HTMLElement>, "color">;