+++ packages/react/src/components/InteractiveList/src/components/InteractiveListStructureSkeleton/index.ts
@@ -1,0 +1,4 @@
+import { InteractiveListStructureSkeleton } from "./InteractiveListStructureSkeleton";
+
+export { InteractiveListStructureSkeleton } from "./InteractiveListStructureSkeleton";
+export default InteractiveListStructureSkeleton;