+++ packages/react/src/components/InteractiveList/src/components/InteractiveListToggleItemSkeleton/index.ts
@@ -1,0 +1,4 @@
+import { InteractiveListToggleItemSkeleton } from "./InteractiveListToggleItemSkeleton";
+
+export { InteractiveListToggleItemSkeleton } from "./InteractiveListToggleItemSkeleton";
+export default InteractiveListToggleItemSkeleton;